requirements.txt
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
bz2file==0.98
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
diskcache==5.6.3
filelock==3.12.2
fsspec==2024.9.0
greenlet==3.0.3
huggingface-hub==0.25.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.4
jmespath==1.0.1
llama_cpp_python==0.3.1
lxml==4.9.3
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parsel==1.8.1
Pillow==10.0.0
pip-autoremove==0.10.0
playwright==1.47.0
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==12.0.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.2
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.10.0
scrapy-rotating-proxies==0.6.2
service-identity==23.1.0
six==1.16.0
soupsieve==2.4.1
tldextract==3.4.4
tqdm==4.66.5
Twisted==22.10.0
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
w3lib==2.1.2
zope.interface==6.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
faiss-cpu==1.8.0.post1
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.8
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langsmith==0.1.130
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.51.0
orjson==3.10.7
packaging==24.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.13.1

